% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songscrape.R
\name{songlist}
\alias{songlist}
\title{songscrape
A clean function that takes the name of an artist and scrapes all songs, lyrics,
album and year details from AZLyrics.com into a clean dataframe. This can be used for
text-mining or whatever you wish.}
\usage{
songlist(artistname)
}
\arguments{
\item{azlyricsname}{Artist name as it appears on the AZlyrics url but in double-quotes}
}
\value{
A dataframe of all lyrics by year, album and song name
}
\description{
songscrape
A clean function that takes the name of an artist and scrapes all songs, lyrics,
album and year details from AZLyrics.com into a clean dataframe. This can be used for
text-mining or whatever you wish.
}
